+++ lib/src/socket_connector.dart
@@ -21,6 +21,8 @@ import 'package:socket_connector/src/types.dart';
 class SocketConnector {
   static const defaultTimeout = Duration(seconds: 30);
 
+  bool gracePeriodPassed = false;
+
   SocketConnector({
     this.verbose = false,
     this.logTraffic = false,
@@ -29,6 +31,7 @@ class SocketConnector {
   }) {
     this.logger = logger ?? stderr;
     Timer(timeout, () {
+      gracePeriodPassed = true;
       if (connections.isEmpty) {
         close();
       }
@@ -106,11 +109,12 @@ class SocketConnector {
         }
       } catch (e) {
         thisSide.authenticated = false;
-        _log('Error while authenticating side ${thisSide.name} : $e');
+        _log('Error while authenticating side ${thisSide.name} : $e',
+            force: true);
       }
     }
     if (!thisSide.authenticated) {
-      _log('Authentication failed on side ${thisSide.name}');
+      _log('Authentication failed on side ${thisSide.name}', force: true);
       _destroySide(thisSide);
       return;
     }
@@ -149,7 +153,7 @@ class SocketConnector {
           _log('stream.onDone on side ${side.name}');
           _destroySide(side);
         }, onError: (error) {
-          _log('stream.onError on side ${side.name}: $error');
+          _log('stream.onError on side ${side.name}: $error', force: true);
           _destroySide(side);
         });
       }
@@ -181,8 +185,9 @@ class SocketConnector {
       if (connectionToRemove != null) {
         connections.remove(connectionToRemove);
         _log(chalk.brightBlue('Removed connection'));
-        if (connections.isEmpty) {
-          _log(chalk.brightBlue('No established connections remain - '
+        if (connections.isEmpty && gracePeriodPassed) {
+          _log(chalk.brightBlue('No established connections remain'
+              ' and grace period has passed - '
               ' will close connector'));
           close();
         }
@@ -214,8 +219,8 @@ class SocketConnector {
     pendingB.clear();
   }
 
-  void _log(String s) {
-    if (verbose) {
+  void _log(String s, {bool force = false}) {
+    if (verbose || force) {
       logger.writeln('${DateTime.now()} | SocketConnector | $s');
     }
   }
@@ -331,6 +336,7 @@ class SocketConnector {
   /// - Creates socket to [portB] on [addressB]
   /// - Relays data between the sockets
   static Future<SocketConnector> socketToSocket({
+    SocketConnector? connector,
     required InternetAddress addressA,
     required int portA,
     required InternetAddress addressB,
@@ -343,7 +349,7 @@ class SocketConnector {
     IOSink? logger,
   }) async {
     IOSink logSink = logger ?? stderr;
-    SocketConnector connector = SocketConnector(
+    connector ??= SocketConnector(
       verbose: verbose,
       logTraffic: logTraffic,
       timeout: timeout,
@@ -374,22 +380,27 @@ class SocketConnector {
   /// - Binds to [portA] on [addressA]
   /// - Listens for a socket connection on [portA] port and joins it to
   ///   the 'B' side
-  ///
   /// - If [portA] is not provided then a port is chosen by the OS.
   /// - [addressA] defaults to [InternetAddress.anyIPv4]
-  static Future<SocketConnector> serverToSocket({
-    /// Defaults to [InternetAddress.anyIPv4]
-    InternetAddress? addressA,
-    int portA = 0,
-    required InternetAddress addressB,
-    required int portB,
-    DataTransformer? transformAtoB,
-    DataTransformer? transformBtoA,
-    bool verbose = false,
-    bool logTraffic = false,
-    Duration timeout = SocketConnector.defaultTimeout,
-    IOSink? logger,
-  }) async {
+  /// - [multi] flag controls whether or not to allow multiple connections
+  ///   to the bound server port [portA]
+  /// - [onConnect] is called when [portA] has got a new connection and a
+  ///   corresponding outbound socket has been created to [addressB]:[portB]
+  static Future<SocketConnector> serverToSocket(
+      {
+      /// Defaults to [InternetAddress.anyIPv4]
+      InternetAddress? addressA,
+      int portA = 0,
+      required InternetAddress addressB,
+      required int portB,
+      DataTransformer? transformAtoB,
+      DataTransformer? transformBtoA,
+      bool verbose = false,
+      bool logTraffic = false,
+      Duration timeout = SocketConnector.defaultTimeout,
+      IOSink? logger,
+      bool multi = false,
+      Function(Socket sideA, Socket sideB)? onConnect}) async {
     IOSink logSink = logger ?? stderr;
     addressA ??= InternetAddress.anyIPv4;
 
@@ -400,18 +411,27 @@ class SocketConnector {
       logger: logSink,
     );
 
+    int connections = 0;
     // bind to a local port for side 'A'
     connector._serverSocketA = await ServerSocket.bind(addressA, portA);
     // listen on the local port and connect the inbound socket
-    connector._serverSocketA?.listen((socket) {
-      Side sideA = Side(socket, true, transformer: transformAtoB);
+    connector._serverSocketA?.listen((sideASocket) async {
+      if (!multi) {
+        unawaited(connector._serverSocketA?.close());
+      }
+      Side sideA = Side(sideASocket, true, transformer: transformAtoB);
       unawaited(connector.handleSingleConnection(sideA));
-    });
 
-    // connect to the side 'B' address and port
-    Socket sideBSocket = await Socket.connect(addressB, portB);
-    Side sideB = Side(sideBSocket, false, transformer: transformBtoA);
-    unawaited(connector.handleSingleConnection(sideB));
+      if (verbose) {
+        logSink.writeln('Making connection ${++connections} to the "B" side');
+      }
+      // connect to the side 'B' address and port
+      Socket sideBSocket = await Socket.connect(addressB, portB);
+      Side sideB = Side(sideBSocket, false, transformer: transformBtoA);
+      unawaited(connector.handleSingleConnection(sideB));
+
+      onConnect?.call(sideASocket, sideBSocket);
+    });
 
     return (connector);
   }